pyproject.toml
[project]
name = "task-manager-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==3.2.2",
    "email-validator>=2.3.0",
    "fastapi>=0.129.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.5.0",
]
